add time in MOD11

+++ Collect/MOD11/DataAccess.py
@@ -78,7 +78,7 @@ def DownloadData(Dir, Startdate, Enddate, latlim, lonlim, TimeStep, Waitbar, cor
 
     # Make directory for the MODIS NDVI data
     Dir = Dir.replace("/", os.sep)
-    output_folder = os.path.join(Dir, 'LST', 'MODIS', '%s' %TimeStepName)
+    output_folder = os.path.join(Dir, 'LST', 'MOD11', '%s' %TimeStepName)
     if not os.path.exists(output_folder):
         os.makedirs(output_folder)
 
@@ -125,33 +125,56 @@ def RetrieveData(Date, args):
     # Argument
     [output_folder, TilesVertical, TilesHorizontal,lonlim, latlim, TimeStep, hdf_library] = args
 
-    # Collect the data from the MODIS webpage and returns the data and lat and long in meters of those tiles
-    try:
-        Collect_data(TilesHorizontal, TilesVertical, Date, output_folder, TimeStep, hdf_library)
-    except:
-        print("Was not able to download the file")
-
-    # Define the output name of the collect data function
-    name_collect = os.path.join(output_folder, 'Merged.tif')
-
-    # Reproject the MODIS product to epsg_to
-    epsg_to ='4326'
-    name_reprojected = RC.reproject_MODIS(name_collect, epsg_to)
-
-    # Clip the data to the users extend
-    data, geo = RC.clip_data(name_reprojected, latlim, lonlim)
-
-    # Save results as Gtiff
     if TimeStep == 8:
-        LSTfileName = os.path.join(output_folder, 'LST_MOD11A2_K_8-daily_' + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '.tif')
+        LSTfileNamePart = os.path.join(output_folder, 'LST_MOD11A2_K_8-daily_' + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '.tif')
     if TimeStep == 1:
-        LSTfileName = os.path.join(output_folder, 'LST_MOD11A1_K_daily_' + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '.tif')
+        LSTfileNamePart = os.path.join(output_folder, 'LST_MOD11A1_K_daily_' + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '.*.tif')
+  
+    filesMOD = glob.glob(LSTfileNamePart)
 
-    DC.Save_as_tiff(name=LSTfileName, data=data, geo=geo, projection='WGS84')
+    if not len(filesMOD) == 1:
 
-    # remove the side products
-    os.remove(os.path.join(output_folder, name_collect))
-    os.remove(os.path.join(output_folder, name_reprojected))
+        # Collect the data from the MODIS webpage and returns the data and lat and long in meters of those tiles
+        try:
+            Collect_data(TilesHorizontal, TilesVertical, Date, output_folder, TimeStep, hdf_library)
+        except:
+            print("Was not able to download the file")
+    
+        # Define the output name of the collect data function
+        name_collect = os.path.join(output_folder, 'Merged.tif')
+    
+        # Reproject the MODIS product to epsg_to
+        epsg_to ='4326'
+        name_reprojected = RC.reproject_MODIS(name_collect, epsg_to)
+    
+        # Clip the data to the users extend
+        data, geo = RC.clip_data(name_reprojected, latlim, lonlim)
+            
+        # Save results as Gtiff
+        if TimeStep == 8:
+            LSTfileName = os.path.join(output_folder, 'LST_MOD11A2_K_8-daily_' + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '.tif')
+        if TimeStep == 1:
+            name_collect_time = os.path.join(output_folder, 'Merged_Time.tif')
+            name_reprojected_time = RC.reproject_MODIS(name_collect_time, epsg_to) 
+            data_time, geo = RC.clip_data(name_reprojected_time, latlim, lonlim)
+            data_time[data_time==25.5] = np.nan
+            data_time_ave = np.nanmean(data_time)
+            try:
+                hour_GMT = int(np.floor(data_time_ave))
+                minutes_GMT = int((data_time_ave - np.floor(data_time_ave))*60)    
+            except:
+                hour_GMT = int(12)
+                minutes_GMT = int(0)
+            LSTfileName = os.path.join(output_folder, 'LST_MOD11A1_K_daily_' + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '.%02d%02d.tif'%(hour_GMT,minutes_GMT))
+            os.remove(name_collect_time)
+            os.remove(name_reprojected_time) 
+    
+        data[data==0.] = -9999
+        DC.Save_as_tiff(name=LSTfileName, data=data, geo=geo, projection='WGS84')
+    
+        # remove the side products
+        os.remove(os.path.join(output_folder, name_collect))
+        os.remove(os.path.join(output_folder, name_reprojected))
 
     return True
 
@@ -226,7 +249,9 @@ def Collect_data(TilesHorizontal,TilesVertical,Date,output_folder, TimeStep, hdf
     sizeX = int((TilesHorizontal[1] - TilesHorizontal[0] + 1) * 1200)
     sizeY = int((TilesVertical[1] - TilesVertical[0] + 1) * 1200)
     DataTot = np.zeros((sizeY, sizeX))
-
+    if TimeStep == 1:    
+        DataTot_Time = np.zeros((sizeY, sizeX))
+        
     # Load accounts
     username, password = WebAccounts.Accounts(Type = 'NASA')
 
@@ -328,9 +353,10 @@ def Collect_data(TilesHorizontal,TilesVertical,Date,output_folder, TimeStep, hdf
                 # Open .hdf only band with NDVI and collect all tiles to one array
                 dataset = gdal.Open(file_name)
                 sdsdict = dataset.GetMetadata('SUBDATASETS')
-                sdslist = [sdsdict[k] for k in sdsdict.keys() if '_1_NAME' in k]
+                sdslist = [sdsdict[k] for k in sdsdict.keys() if (('SUBDATASET_1_NAME') in k or ('SUBDATASET_3_NAME') in k)]
                 sds = []
-
+                sds_time = []
+                
                 for n in sdslist:
                     sds.append(gdal.Open(n))
                     full_layer = [i for i in sdslist if 'LST_Day_1km' in i]
@@ -345,6 +371,15 @@ def Collect_data(TilesHorizontal,TilesVertical,Date,output_folder, TimeStep, hdf
                     countYdata = (TilesVertical[1] - TilesVertical[0] + 2) - countY
                     DataTot[int((countYdata - 1) * 1200):int(countYdata * 1200), int((countX - 1) * 1200):int(countX * 1200)]=data * 0.02
                 del data
+ 
+                if TimeStep == 1:
+                    full_layer_time = [i for i in sdslist if 'Day_view_time' in i]
+                    idx_time = sdslist.index(full_layer_time[0])
+                    sds_time.append(gdal.Open(sdslist[idx_time]))                
+                    data_time = sds_time[0].ReadAsArray()
+                    DataTot_Time[int((countYdata - 1) * 1200):int(countYdata * 1200), int((countX - 1) * 1200):int(countX * 1200)]=data_time * 0.1
+                    del data_time
+
 
             # if the tile not exists or cannot be opened, create a nan array with the right projection
             except:
@@ -359,8 +394,14 @@ def Collect_data(TilesHorizontal,TilesVertical,Date,output_folder, TimeStep, hdf
                 countYdata=(TilesVertical[1] - TilesVertical[0] + 2) - countY
                 DataTot[(countYdata - 1) * 1200:countYdata * 1200,(countX - 1) * 1200:countX * 4800] = data * 0.02
                 DataTot[DataTot < 1] = -9999
+                
+                if TimeStep == 1:
+                    data_time=np.ones((1200,1200)) * (-9999/0.1)
+                    DataTot_Time[(countYdata - 1) * 1200:countYdata * 1200,(countX - 1) * 1200:countX * 4800] = data_time * 0.1
+                    
 
     # Make geotiff file
+    DataTot[DataTot < 1] = -9999    
     name2 = os.path.join(output_folder, 'Merged.tif')
     driver = gdal.GetDriverByName("GTiff")
     dst_ds = driver.Create(name2, DataTot.shape[1], DataTot.shape[0], 1, gdal.GDT_Float32, ['COMPRESS=LZW'])
@@ -379,4 +420,27 @@ def Collect_data(TilesHorizontal,TilesVertical,Date,output_folder, TimeStep, hdf
     dst_ds.GetRasterBand(1).WriteArray(DataTot)
     dst_ds = None
     sds = None
+
+    if TimeStep == 1:    
+        DataTot_Time[np.logical_and(DataTot_Time < 0., DataTot_Time>24.)] = -9999
+        # Make geotiff file of the time data
+        name_out_time = os.path.join(output_folder, 'Merged_Time.tif')
+        driver = gdal.GetDriverByName("GTiff")
+        dst_ds = driver.Create(name_out_time, DataTot_Time.shape[1], DataTot_Time.shape[0], 1, gdal.GDT_Float32, ['COMPRESS=LZW'])
+        try:
+             dst_ds.SetProjection(proj)
+        except:
+            proj='PROJCS["unnamed",GEOGCS["Unknown datum based upon the custom spheroid",DATUM["Not specified (based on custom spheroid)",SPHEROID["Custom spheroid",6371007.181,0]],PRIMEM["Greenwich",0],UNIT["degree",0.0174532925199433]],PROJECTION["Sinusoidal"],PARAMETER["longitude_of_center",0],PARAMETER["false_easting",0],PARAMETER["false_northing",0],UNIT["Meter",1]]'
+            x1 = (TilesHorizontal[0] - 18) * 1200 * Distance
+            x4 = (TilesVertical[0] - 9) * 1200 * -1 * Distance
+            geo = [x1, Distance, 0.0, x4, 0.0, -Distance]
+            geo_t = tuple(geo)
+            dst_ds.SetProjection(proj)
+    
+        dst_ds.GetRasterBand(1).SetNoDataValue(-9999)
+        dst_ds.SetGeoTransform(geo_t)
+        dst_ds.GetRasterBand(1).WriteArray(DataTot_Time)
+        dst_ds = None
+        sds = None        
+    
     return()
